refactor: move handbook components in to Docs subfolder
This commit moves components that were previously in the templates
folder under the Handbook section into their own Docs folder in the
components section. This is to make creating new layouts that are based
off the documentation layout easier to create
