+++ src/components/Docs/Footer.js
